style(#916): cargo fmt + clippy unnecessary_cast on the zero-fill helper
`((rd_bits as u16) << 8) as u16` — the outer cast is a no-op (rust-1.96
clippy::unnecessary_cast, -D warnings in CI).

Gate sweep by REAL exit code, not piped output:
  cargo fmt --check                                    exit 0
  cargo clippy --workspace --all-targets -D warnings   exit 0
  cargo test --workspace                               exit 0
  frozen_codegen_bytes (10 anchors)                    exit 0, BYTE-IDENTICAL
  estimator_encoder_agreement (#498)                   exit 0
  scripts/oracle_wiring_check.py                       exit 0
  scripts/claim_check.py                               exit 0, 38/38

NO frozen fixture moved. The fix is confined to `rd >= R8` and none of the four
anchors (control_step, flight_seam, flight_seam_flat, signed_div_const) has a
high-register i64 zero-fill destination, so their .text sha256 pins are unchanged
— nothing to re-pin, and no re-pin to justify.

Refs #916

+++ crates/synth-backend/src/arm_encoder.rs
@@ -8282,7 +8282,7 @@ fn emit_thumb_zero_fill(bytes: &mut Vec<u8>, rd_bits: u32) {
         bytes.extend_from_slice(&movs.to_le_bytes());
     } else {
         bytes.extend_from_slice(&0xF04Fu16.to_le_bytes());
-        bytes.extend_from_slice(&(((rd_bits as u16) << 8) as u16).to_le_bytes());
+        bytes.extend_from_slice(&((rd_bits as u16) << 8).to_le_bytes());
     }
 }
